Rewrote Region as a ranged based data structure, removed scene system
First, I've removed any vestige of the scene system, since it was just
extra code bloating the project at this point.

Map Utility is currently disabled, don't use this version.

I've created a class called Tile, which in addition to the value of the
tile, also carries an x & y value which can be useful down the road. It
also helps by having an iterator datatype.

Region is the biggest focus of this commit. It now uses a ranged based
system for storing the tiles, rather than a "***" for storing 3D arrays.
The current map format is awkward because of the switch, and will probably
be changed in the future.